+++ net/minecraft/client/render/BufferRenderer.java
@@ -119,7 +119,7 @@ public class BufferRenderer {
         }
         if (shader.screenSize != null) {
             Window window = MinecraftClient.getInstance().getWindow();
-            shader.screenSize.set(window.getFramebufferWidth(), window.getFramebufferHeight());
+            shader.screenSize.set((float)window.getFramebufferWidth(), (float)window.getFramebufferHeight());
         }
         if (shader.lineWidth != null && (drawMode == VertexFormat.DrawMode.LINES || drawMode == VertexFormat.DrawMode.LINE_STRIP)) {
             shader.lineWidth.set(RenderSystem.getShaderLineWidth());
